+++ src/servers/model/utils/fetch-retry.ts
@@ -1,11 +1,11 @@
 /**
- * Copyright (c) 2018-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Adam Midlik <midlik@gmail.com>
+ * @author Paul Pillot <paul.pillot@tandemai.com>
  */
 
-import fetch, { AbortError, Response } from 'node-fetch';
 import { retryIf } from '../../../mol-util/retry-if';
 import { downloadGs } from '../../common/google-cloud-storage';
 
@@ -36,13 +36,13 @@ export async function fetchRetry(url: string, timeout: number, retryCount: numbe
 }
 
 /** Like `fetch` but supports Google Cloud Storage (gs://) protocol. */
-export function wrapFetch(url: string, init?: fetch.RequestInit): Promise<Response> {
+export function wrapFetch(url: string, init?: RequestInit): Promise<Response> {
     if (url.startsWith('gs://')) return fetchGS(url, init);
     else return fetch(url, init);
 }
 
-async function fetchGS(url: string, init?: fetch.RequestInit): Promise<Response> {
-    if (init?.signal?.aborted) throw new AbortError('The user aborted a request.');
+async function fetchGS(url: string, init?: RequestInit): Promise<Response> {
+    if (init?.signal?.aborted) throw new Error('The user aborted a request.');
     const data = await downloadGs(url);
     return new Response(data, init);
 }